Fix infiniteLoop bug, refs #293
After adding many or very wide images the infiniteLoop would
break and just "rewind" back to the beginning or end of the slider.

This was due to the slider's wrapper div not being wide enough.

So instead of setting an arbitrary percentage, all the children
elements of the slider are measured and the wrapper width
is set accordingly.