Add public method to set active index
For scenarios where changing slides on one slider will change slides on another on the same page.

There is a bug somewhere either here on in my code that uses this library whereby jumping back to the start from any other position does not move the slider because it thinks it is already at slide 0.  This tricks the slider into moving.

Until I figure out what is truly going on, this is a quick and dirty fix.