Label Changes
Taking label from the label attribute on <img> tag. E.g: <img src ='img.jpeg' label='<p>This is a label</ p>' title="This is a label" /> 

Because when inserting html in title, the label displayed the html code, but in img [title = attr] displayed the content in html too. That way you can use html code in img [attr = label] and titles separately.