Set pointer capture on the right target
Set the pointer capture on the slider itself only
when we are in the swiping mode. The ensures not only
clicks would work on elements such as links within the
slider if user hasn't moved the pointer, but also
disable activating and navigating to the links within
the slider if the user has started the slider move on
an embedded link within the slider.